requirements.txt
grpcio==1.67.0
grpcio-tools==1.67.0
requests==2.32.3
pymongo==4.8.0
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
fastapi==0.115.2
uvicorn==0.30.6
pydantic-settings==2.6.0
